feat: add queue
Add a message queue (using the internal Python queue class) for communication between subliminal and the UI. To summarize, the new features are:
- Last 1000 log lines are kept in memory (this app is designed with a Raspbery Pi in mind, so I want to minimize writes to the SD card)
- A progress bar has been implemented
- As the UI is now aware of when the process has finished, the action buttons are re-enabled automatically.

It's starting to resemble a functional app. Next steps will be:
- Adding tabs to the UI (it's a bit crowded right now)
- Letting users choose the providers and the order of preference by dragging them in a list
- Adding a notification system
- Maybe, just maybe, adding a feature to manually choose subtitles

Keep in mind that I intend this app to be lightweight, so I want to minimize the amount of libraries I'm using. Sometimes less is better.